requirements.txt
Flask>=2.2
Pillow>=9.0
requests>=2.28
openai>=1.0.0
python-dotenv>=1.0
